requirements.txt
fastapi
uvicorn
google-generativeai
pypdf
python-pptx
python-multipart 